+++ src/util/pp.mli
@@ -121,10 +121,10 @@ val string_of_wnl : 'a pp -> 'a -> string
 val wnl : formatter -> unit
 
 val sprintf :
-  ('b,  formatter, unit, string) Pervasives.format4 -> 'b
+  ('b,  formatter, unit, string) Stdlib.format4 -> 'b
 
 val sprintf_wnl :
-  ('b,  formatter, unit, string) Pervasives.format4 -> 'b
+  ('b,  formatter, unit, string) Stdlib.format4 -> 'b
 
 val html_char : char pp
 val html_string : string pp