+++ src/util/wstdlib.ml
@@ -13,7 +13,7 @@
 
 module Int = struct
   type t = int
-  let compare (x : int) y = Pervasives.compare x y [@ocaml.warning "-3"]
+  let compare (x : int) y = Stdlib.compare x y
   let equal (x : int) y = x = y
   let hash  (x : int) = x
 end
@@ -33,7 +33,7 @@ end)
 
 module Float = struct
   type t = float
-  let compare (x : float) y = Pervasives.compare x y [@ocaml.warning "-3"]
+  let compare (x : float) y = Stdlib.compare x y
   let equal (x : float) y = x = y
   let hash  (x : float) = Exthtbl.hash x
 end