+++ CHANGES.md
@@ -29,6 +29,9 @@ Provers
   * support for CVC5 1.0.0 (released April 6, 2022)
   * support for Alt-Ergo 2.4.2 (released August 1, 2022)
 
+Miscellaneous
+  * OCaml >= 4.08 is now required
+
 Version 1.5.1, September 12, 2022
 ---------------------------------